# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Movie {
  id: number;
  title: string;
  year: string;
  imdbId?: string;
  plot?: string;
  imdbRating?: string;
  genre?: string;
  director?: string;
  writer?: string;
  actors?: string;
  poster?: string;
  folderName: string;
  folderPath: string;
  notFound: boolean;
  watched: boolean;
  userRating?: number;
  notes?: string;
  createdAt: string;
}

export type MovieStatsGenreBreakdownItem = {
  genre: string;
  count: number;
};

export interface MovieStats {
  total: number;
  watched: number;
  avgRating: number;
  genreBreakdown: MovieStatsGenreBreakdownItem[];
  topRated: Movie[];
  recentlyAdded: Movie[];
}

export type ScanResultDetailsItemStatus =
  (typeof ScanResultDetailsItemStatus)[keyof typeof ScanResultDetailsItemStatus];

export const ScanResultDetailsItemStatus = {
  added: "added",
  skipped: "skipped",
  failed: "failed",
  not_found: "not_found",
} as const;

export type ScanResultDetailsItem = {
  folderName: string;
  title: string;
  status: ScanResultDetailsItemStatus;
  reason?: string;
};

export interface ScanResult {
  added: number;
  skipped: number;
  failed: number;
  notFound: number;
  total: number;
  details: ScanResultDetailsItem[];
}

export interface MovieUpdatePayload {
  watched?: boolean;
  userRating?: number;
  notes?: string;
}

export interface SubtitleFile {
  name: string;
  path: string;
}

export interface MoviePlayPayload {
  subtitlePath?: string;
}

export interface PlayResponse {
  success: boolean;
  message: string;
}

export interface OpenFolderResponse {
  success: boolean;
}

export interface ErrorResponse {
  error: string;
}

export interface DeleteResponse {
  success: boolean;
}

export type ListMoviesParams = {
  search?: string;
  genre?: string;
  director?: string;
  actor?: string;
  year?: string;
  minRating?: string;
  watched?: string;
  sort?: ListMoviesSort;
};

export type ListMoviesSort =
  (typeof ListMoviesSort)[keyof typeof ListMoviesSort];

export const ListMoviesSort = {
  title: "title",
  year: "year",
  rating: "rating",
  added: "added",
} as const;