# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: movies
    description: Movie library operations
  - name: library
    description: Library scan and system operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /movies:
    get:
      operationId: listMovies
      tags: [movies]
      summary: List all movies in the library
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: genre
          in: query
          required: false
          schema:
            type: string
        - name: director
          in: query
          required: false
          schema:
            type: string
        - name: actor
          in: query
          required: false
          schema:
            type: string
        - name: year
          in: query
          required: false
          schema:
            type: string
        - name: minRating
          in: query
          required: false
          schema:
            type: string
        - name: watched
          in: query
          required: false
          schema:
            type: string
        - name: sort
          in: query
          required: false
          schema:
            type: string
            enum: [title, year, rating, added]
      responses:
        "200":
          description: List of movies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Movie"

  /movies/stats:
    get:
      operationId: getMovieStats
      tags: [movies]
      summary: Get library statistics
      responses:
        "200":
          description: Library stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MovieStats"

  /movies/scan:
    post:
      operationId: scanLibrary
      tags: [library]
      summary: Scan the movies folder and fetch IMDB info for new movies
      responses:
        "200":
          description: Scan results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanResult"

  /movies/{id}:
    get:
      operationId: getMovie
      tags: [movies]
      summary: Get a movie by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Movie details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Movie"
        "404":
          description: Movie not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteMovie
      tags: [movies]
      summary: Remove a movie from the library
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Movie deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /movies/{id}/update:
    patch:
      operationId: updateMovie
      tags: [movies]
      summary: Update watched status, personal rating, or notes
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MovieUpdatePayload"
      responses:
        "200":
          description: Updated movie
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Movie"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /movies/{id}/subtitles:
    get:
      operationId: getMovieSubtitles
      tags: [movies]
      summary: List subtitle files available for a movie
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of subtitle files
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SubtitleFile"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /movies/{id}/play:
    post:
      operationId: playMovie
      tags: [movies]
      summary: Launch VLC to play the movie with optional subtitle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MoviePlayPayload"
      responses:
        "200":
          description: VLC launched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /movies/{id}/open-folder:
    post:
      operationId: openMovieFolder
      tags: [movies]
      summary: Open the movie folder in Windows Explorer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Folder opened
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenFolderResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Movie:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        year:
          type: string
        imdbId:
          type: string
        plot:
          type: string
        imdbRating:
          type: string
        genre:
          type: string
        director:
          type: string
        writer:
          type: string
        actors:
          type: string
        poster:
          type: string
        folderName:
          type: string
        folderPath:
          type: string
        notFound:
          type: boolean
        watched:
          type: boolean
        userRating:
          type: integer
        notes:
          type: string
        createdAt:
          type: string
      required:
        - id
        - title
        - year
        - folderName
        - folderPath
        - notFound
        - watched
        - createdAt

    MovieStats:
      type: object
      properties:
        total:
          type: integer
        watched:
          type: integer
        avgRating:
          type: number
        genreBreakdown:
          type: array
          items:
            type: object
            properties:
              genre:
                type: string
              count:
                type: integer
            required:
              - genre
              - count
        topRated:
          type: array
          items:
            $ref: "#/components/schemas/Movie"
        recentlyAdded:
          type: array
          items:
            $ref: "#/components/schemas/Movie"
      required:
        - total
        - watched
        - avgRating
        - genreBreakdown
        - topRated
        - recentlyAdded

    ScanResult:
      type: object
      properties:
        added:
          type: integer
        skipped:
          type: integer
        failed:
          type: integer
        notFound:
          type: integer
        total:
          type: integer
        details:
          type: array
          items:
            type: object
            properties:
              folderName:
                type: string
              title:
                type: string
              status:
                type: string
                enum: [added, skipped, failed, not_found]
              reason:
                type: string
            required:
              - folderName
              - title
              - status
      required:
        - added
        - skipped
        - failed
        - notFound
        - total
        - details

    MovieUpdatePayload:
      type: object
      properties:
        watched:
          type: boolean
        userRating:
          type: integer
        notes:
          type: string

    SubtitleFile:
      type: object
      properties:
        name:
          type: string
        path:
          type: string
      required:
        - name
        - path

    MoviePlayPayload:
      type: object
      properties:
        subtitlePath:
          type: string

    PlayResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message

    OpenFolderResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    DeleteResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success