# lib/api-zod/src/generated/types/movie.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Movie {
  id: number;
  title: string;
  year: string;
  imdbId?: string;
  plot?: string;
  imdbRating?: string;
  genre?: string;
  director?: string;
  writer?: string;
  actors?: string;
  poster?: string;
  folderName: string;
  folderPath: string;
  notFound: boolean;
  watched: boolean;
  userRating?: number;
  notes?: string;
  createdAt: string;
}