# lib/api-zod/src/generated/types/movieStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Movie } from "./movie";
import type { MovieStatsGenreBreakdownItem } from "./movieStatsGenreBreakdownItem";

export interface MovieStats {
  total: number;
  watched: number;
  avgRating: number;
  genreBreakdown: MovieStatsGenreBreakdownItem[];
  topRated: Movie[];
  recentlyAdded: Movie[];
}