# lib/api-zod/src/generated/types/scanResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ScanResultDetailsItem } from "./scanResultDetailsItem";

export interface ScanResult {
  added: number;
  skipped: number;
  failed: number;
  notFound: number;
  total: number;
  details: ScanResultDetailsItem[];
}